Refactor patrol area class to new file
