+++ src/openrct2/entity/Staff.cpp
@@ -40,6 +40,7 @@
 #include "../world/Scenery.h"
 #include "../world/SmallScenery.h"
 #include "../world/Surface.h"
+#include "PatrolArea.h"
 #include "Peep.h"
 
 #include <algorithm>
@@ -61,127 +62,9 @@ const rct_string_id StaffCostumeNames[] = {
 };
 // clang-format on
 
-uint16_t gStaffDrawPatrolAreas;
 colour_t gStaffHandymanColour;
 colour_t gStaffMechanicColour;
 colour_t gStaffSecurityColour;
-
-static PatrolArea _mergedPatrolAreas[EnumValue(StaffType::Count)];
-
-static bool CompareTileCoordsXY(const TileCoordsXY& lhs, const TileCoordsXY& rhs)
-{
-    if (lhs.y == rhs.y)
-        return lhs.x < rhs.x;
-    return lhs.y < rhs.y;
-}
-
-const PatrolArea::Cell* PatrolArea::GetCell(TileCoordsXY pos) const
-{
-    return const_cast<PatrolArea*>(this)->GetCell(pos);
-}
-
-PatrolArea::Cell* PatrolArea::GetCell(TileCoordsXY pos)
-{
-    auto areaPos = TileCoordsXY(pos.x / Cell::Width, pos.y / Cell::Height);
-    if (areaPos.x < 0 || areaPos.x >= CellColumns || areaPos.y < 0 || areaPos.y >= CellRows)
-        return nullptr;
-
-    auto& area = Areas[(areaPos.y * CellColumns) + areaPos.x];
-    return &area;
-}
-
-bool PatrolArea::IsEmpty() const
-{
-    return TileCount == 0;
-}
-
-void PatrolArea::Clear()
-{
-    for (auto& area : Areas)
-    {
-        area.SortedTiles.clear();
-    }
-}
-
-bool PatrolArea::Get(TileCoordsXY pos) const
-{
-    auto* area = GetCell(pos);
-    if (area == nullptr)
-        return false;
-
-    auto it = std::lower_bound(area->SortedTiles.begin(), area->SortedTiles.end(), pos, CompareTileCoordsXY);
-    auto found = it != area->SortedTiles.end() && *it == pos;
-    return found;
-}
-
-bool PatrolArea::Get(CoordsXY pos) const
-{
-    return Get(TileCoordsXY(pos));
-}
-
-void PatrolArea::Set(TileCoordsXY pos, bool value)
-{
-    auto* area = GetCell(pos);
-    if (area == nullptr)
-        return;
-
-    auto it = std::lower_bound(area->SortedTiles.begin(), area->SortedTiles.end(), pos, CompareTileCoordsXY);
-    auto found = it != area->SortedTiles.end() && *it == pos;
-
-    if (!found && value)
-    {
-        area->SortedTiles.insert(it, pos);
-        TileCount++;
-    }
-    else if (found && !value)
-    {
-        area->SortedTiles.erase(it);
-        assert(TileCount != 0);
-        TileCount--;
-    }
-}
-
-void PatrolArea::Set(CoordsXY pos, bool value)
-{
-    Set(TileCoordsXY(pos), value);
-}
-
-void PatrolArea::Union(const PatrolArea& other)
-{
-    for (size_t i = 0; i < Areas.size(); i++)
-    {
-        for (const auto& pos : other.Areas[i].SortedTiles)
-        {
-            Set(pos, true);
-        }
-    }
-}
-
-void PatrolArea::Union(const std::vector<TileCoordsXY>& other)
-{
-    for (const auto& pos : other)
-    {
-        Set(pos, true);
-    }
-}
-
-std::vector<TileCoordsXY> PatrolArea::ToVector() const
-{
-    std::vector<TileCoordsXY> result;
-    for (const auto& area : Areas)
-    {
-        for (const auto& pos : area.SortedTiles)
-        {
-            result.push_back(pos);
-        }
-    }
-    return result;
-}
-
-const PatrolArea& GetMergedPatrolArea(const StaffType type)
-{
-    return _mergedPatrolAreas[EnumValue(type)];
-}
 
 // Maximum manhattan distance that litter can be for a handyman to seek to it
 const uint16_t MAX_LITTER_DISTANCE = 3 * COORDS_XY_STEP;
@@ -189,40 +72,6 @@ const uint16_t MAX_LITTER_DISTANCE = 3 * COORDS_XY_STEP;
 template<> bool EntityBase::Is<Staff>() const
 {
     return Type == EntityType::Staff;
-}
-
-/**
- *
- *  rct2: 0x006BD3A4
- */
-void staff_reset_modes()
-{
-    staff_update_greyed_patrol_areas();
-}
-
-/**
- *
- *  rct2: 0x006C0C3F
- */
-void staff_update_greyed_patrol_areas()
-{
-    for (int32_t staffType = 0; staffType < EnumValue(StaffType::Count); ++staffType)
-    {
-        // Reset all of the merged data for the type.
-        auto& mergedArea = _mergedPatrolAreas[staffType];
-        mergedArea.Clear();
-
-        for (auto staff : EntityList<Staff>())
-        {
-            if (EnumValue(staff->AssignedStaffType) != staffType)
-                continue;
-
-            if (staff->PatrolInfo == nullptr)
-                return;
-
-            mergedArea.Union(*staff->PatrolInfo);
-        }
-    }
 }
 
 /**
@@ -431,11 +280,6 @@ bool Staff::IsPatrolAreaSet(const CoordsXY& coords) const
         return PatrolInfo->Get(coords);
     }
     return false;
-}
-
-bool staff_is_patrol_area_set_for_type(StaffType type, const CoordsXY& coords)
-{
-    return _mergedPatrolAreas[EnumValue(type)].Get(coords);
 }
 
 void Staff::SetPatrolArea(const CoordsXY& coords, bool value)
