+++ src/openrct2/entity/Staff.h
@@ -14,45 +14,7 @@
 #include "Peep.h"
 
 class DataSerialiser;
-
-// The number of elements in the gStaffPatrolAreas array per staff member. Every bit in the array represents a 4x4 square.
-// Right now, it's a 32-bit array like in RCT2. 32 * 128 = 4096 bits, which is also the number of 4x4 squares on a 256x256 map.
-constexpr size_t STAFF_PATROL_AREA_BLOCKS_PER_LINE = MAXIMUM_MAP_SIZE_TECHNICAL / 4;
-constexpr size_t STAFF_PATROL_AREA_SIZE = (STAFF_PATROL_AREA_BLOCKS_PER_LINE * STAFF_PATROL_AREA_BLOCKS_PER_LINE) / 32;
-
-struct PatrolArea
-{
-private:
-    struct Cell
-    {
-        static constexpr auto Width = 64;
-        static constexpr auto Height = 64;
-        static constexpr auto NumTiles = Width * Height;
-
-        std::vector<TileCoordsXY> SortedTiles;
-    };
-
-    static constexpr auto CellColumns = (MAXIMUM_MAP_SIZE_TECHNICAL + (Cell::Width - 1)) / Cell::Width;
-    static constexpr auto CellRows = (MAXIMUM_MAP_SIZE_TECHNICAL + (Cell::Height - 1)) / Cell::Height;
-    static constexpr auto NumCells = CellColumns * CellRows;
-
-    std::array<Cell, NumCells> Areas;
-    size_t TileCount{};
-
-    const Cell* GetCell(TileCoordsXY pos) const;
-    Cell* GetCell(TileCoordsXY pos);
-
-public:
-    bool IsEmpty() const;
-    void Clear();
-    bool Get(TileCoordsXY pos) const;
-    bool Get(CoordsXY pos) const;
-    void Set(TileCoordsXY pos, bool value);
-    void Set(CoordsXY pos, bool value);
-    void Union(const PatrolArea& other);
-    void Union(const std::vector<TileCoordsXY>& other);
-    std::vector<TileCoordsXY> ToVector() const;
-};
+class PatrolArea;
 
 struct Staff : Peep
 {
@@ -181,14 +143,10 @@ enum class EntertainerCostume : uint8_t
 
 extern const rct_string_id StaffCostumeNames[static_cast<uint8_t>(EntertainerCostume::Count)];
 
-extern uint16_t gStaffDrawPatrolAreas;
 extern colour_t gStaffHandymanColour;
 extern colour_t gStaffMechanicColour;
 extern colour_t gStaffSecurityColour;
 
-void staff_reset_modes();
-void staff_update_greyed_patrol_areas();
-bool staff_is_patrol_area_set_for_type(StaffType type, const CoordsXY& coords);
 colour_t staff_get_colour(StaffType staffType);
 bool staff_set_colour(StaffType staffType, colour_t value);
 uint32_t staff_get_available_entertainer_costumes();
